Changed the login function to take the username and password from the Http Header
The login function now takes in a username and password from the Http Header instead of a LoginRequest. This change was made because the Authentication Filter can only check values from the header, and it didn't seem like a good idea to send the username and password in both the header and the body at the same time...

So from now on, all authentication details (Username+password, or Jwt tokens), will be provided in the header.